Move cli to a directory
