+++ bin/index.cjs
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 
-require('../lib/cjs/cli.js');
+require("../lib/cjs/cli/cli.js");
