+++ src/history/index.ts
@@ -3,7 +3,7 @@ import {resolveHistoryLocation} from "./location.js";
 import {LocalHistoryProvider} from "./local.js";
 import {getGaCacheHistoryProvider} from "./gaCache.js";
 import {IHistoryProvider} from "./provider.js";
-import {optionsDefault} from "../options.js";
+import {optionsDefault} from "../cli/options.js";
 import {S3HistoryProvider} from "./s3.js";
 export {resolveHistoryLocation};
